don't embed the SVG inside the .shape, link to it
Dia doesn't seem to support complex SVG in the .shape? Or is the problem
elsewhere?
Anyway, just xlinking to the SVG file yields better results, and avoids
parsing the SVG ourselves, so it may be better.
Rely on ImageMagick to get the aspect ratio and dimensions then. Dia
supports abstract coordinates (we'll max them to 64x64).